ffmpeg: Fix buzzing in resampled audio output.
The resampler was not setting the expected number of samples per output
frame. The AAC encoder expects a certain number of samples per frame,
and was presumably "filling in" the missing ssamples, which led to the
buzzing. What we wanted was for the resampler to buffer until it had
enough samples to emit a full frame for the encoder.

For example, going from 48khz -> 44.1khz results in fewer samples, so
"shorter" outputs for a given input frame. The encoder was apparently
padding this up to its expected number of 1024.