ffmpeg: Remove one-thread-per-segment restriction.
This change allows ffmpeg to use as many threads as it can for
the encoding process. The renditions for each frame are still
processed sequentially, but each rendition can now use multiple
threads.

This does make it more difficult to characterize the overall behavior
or capacity of a node, but it is necessary to go fast for well-tuned
nodes.