ffmpeg: Drain queued frames from the filter and encoder.
This allows for better resource management.

Multiple frames may be output for a given filter call, such as
upsampling the frame rate. This led to running out of
surfaces in GPU-encoding mode, and steadily increasing memory usage
in software-encoding mode.

Likewise, GOP structure may result in the encoder having several frames ready
to be returned after being fed earlier frames with no immediate result.

The transcode loop is reworked to:

0. Decode a frame
1. Feed the filter graph with the decoded frame
1a. Feed and drain the encoder if no filter graph is configured
2. Drain the filter graph as much as possible
3. Feed and drain the encoder for each frame received from the filter
4. Mux and output the resulting packets.