ffmpeg: Plug memory leak in filtergraph init.
The in-out structs are allocated but not freed if the function exits
early, such as if decoding is not required.

Make any early exits jump to the cleanup routine to avoid these errors
in the future. Also move the allocations so they don't occur if an early
exit is needed.

Make the audio filter more symmetrical with respect to video.