ffmpeg: Add Format shorthand to VideoProfile.
Initially populate with mpegts and mp4. This can be used in
place of manually specifying muxer options.

If a format is not specified, the richer Muxer transcoding options
are looked at. The default is no format.

If both VideoProfile.Format and TranscodingOptions.Muxer are used,
then VideoProfile.Format takes precedence.