ffmpeg: Fix decoder flush for short segments
Earlier we relied on getting a sentinel (pts -1) packet back to know
when the decoder is flushed. This does not work for short segments as
the decoder buffers packets internally.

Now we maintain the difference between packets sent - frames received
and mark decoder as flushed when this becomes 0 (usually).

Sometimes the decoder might not return all frames (see #155), so we give
up after sending SENTINEL_MAX packets and not getting any valid frame
back.