[WIP] ffmpeg: use first_pkt dts when NULL flushing
The segment order comparing relies on last_frame->pkt_dts, which in turn
relies on the packet that triggered decoding of the frame [1]

When we use NULL flushing the pkt_dts for the returned frame is thus set
to 0. To match the sentinel flushing workflow, manually setting it to
first_pkt->dts here.

[1]: https://ffmpeg.org/doxygen/trunk/structAVFrame.html#aa52951f35ec9e303d3dfeb4b3e44248a