stream_index instead of pkt usage
Removed usage of pkt for a stream_index, because it was used so
even when flushing. And the problem is, when flushing there is
no pkt - previous instance was used.

Also flushing was reworked. It was previously done "twice":
once in main loop, and once after that loop ended. Because I
found flushing in the loop problematic (for one, it was limited
to one encoder, whereas output may have two of them), I only
left after the loop flush.