Fixed packet clone bug
Several tests were failing because of packet clone bug - and
rightly so. In ffmpeg packet given for muxing is consumed by the
muxer, so if it is to be used later, it has to be cloned. Problem
was, I was cloning the packet (with intention of muxing it) but
still passing original one to the muxer. This in turn caused
original packet to be consumed, and when that packet was given
to decoder, it caused end of file errors.

Also some not needed code was removed.