transcode2() error-tolerant
I was reviewing transcode2() source vs old transcode() and found
out that in old code errors in decode loop were "tolerated". What
I mean by that is that such error was interrupting the decode loop
but the flushing of the filters would still be performed.

transcode2() was modified accordingly to replicate previous code's
behaviour.