StreamBuffer for input
Alternative input system for LPMS. The idea is to allow using of
FFmpeg demultiplexers with any input mechanism, such as network
packet delivery. In order to do that we have to provide custom
AVIOContext with custom read() and seek() operations. This also
means that synchronization is needed because we will have to
perform input and transcoding in different threads (for example
demux - which is a part of transcode - may want more data which
is not available yet, then transcode should block until input
thread delivers more data)