Test_ResolutionClamp fixed
Previous commit changed the resolution clamping approach: instead
of changing every resolution given, change only resolutions that
are not acceptable to the encoder. To make tests pass properly,
one thing needed to change. Previously, when 600x300 resolution was
given to the encoder with 700x500 limits, it got changed to the
250x500. While that was an expected behaviour, I don't think it is
a proper one - why change something that encoder is fully willing
to accept? So the test was modified to be OK, with original
600x300