Fix for TestTranscoder_Portrait
Previous changes caused limits from the array in ffmpeg.go to be
strictly applied, i.e. if the resolution was fitting the limits,
no change was done. This uncovered the fact that the limits
themselves were bad - on my machine NVenc refused to work with the
message that the frame dimensions is too low.

Changing the minimum resolution to 146x145 fixed the problem. Note
that is is quite likely that the limits should be different for
h265 as well, but on my machine all works.

Proper approach for this problem would be something along the line
of dynamic discovery of these limits (via Nvidia APIs) and then
applying true local HW limits, not hardcoded ones.