ffmpeg: implement looping
So -stream_loop doesn't work when provided as a
demuxer option, as it's implemented in https://github.com/FFmpeg/FFMpeg/blob/master/fftools/ffmpeg_demux.c
and we handle our own demuxing. Thankfully, this
seemed to be relatively easy to port over using
the preexisting work that we have for handling
discontinuity. Lightly tested.